# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShopMate AaaS Enterprise Python AI Engine",
    "description": "Enterprise Python FastAPI backend powering Database-backed 12-stage RAG, Multi-step Agent Runtime, and E-commerce Tools.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/db/status": {
      "get": {
        "summary": "Get Db Status",
        "description": "Returns the live status of the Enterprise Database (Protected: Admin/Service Token Required).",
        "operationId": "get_db_status_api_v1_db_status_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_id}/chat": {
      "post": {
        "summary": "Chat Agent",
        "description": "Executes a full multi-step agent reasoning cycle with 12-stage RAG and tools.\nWorkspace ID is strictly derived from the verified service JWT.",
        "operationId": "chat_agent_api_v1_agents__agent_id__chat_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__models__ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__models__ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_id}/chat/stream": {
      "post": {
        "summary": "Chat Agent Stream",
        "operationId": "chat_agent_stream_api_v1_agents__agent_id__chat_stream_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__models__ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rag/query": {
      "post": {
        "summary": "Query Rag Pipeline",
        "operationId": "query_rag_pipeline_api_v1_rag_query_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RAGQueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/ingest": {
      "post": {
        "summary": "Ingest Knowledge Endpoint",
        "operationId": "ingest_knowledge_endpoint_api_v1_knowledge_ingest_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeIngestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_number}": {
      "get": {
        "summary": "Get Order Endpoint",
        "operationId": "get_order_endpoint_api_v1_orders__order_number__get",
        "parameters": [
          {
            "name": "order_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Number"
            }
          },
          {
            "name": "customer_email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Customer email for verification",
              "title": "Customer Email"
            },
            "description": "Customer email for verification"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Citation": {
        "properties": {
          "document_name": {
            "type": "string",
            "title": "Document Name"
          },
          "chunk_text": {
            "type": "string",
            "title": "Chunk Text"
          },
          "relevance_score": {
            "type": "number",
            "title": "Relevance Score"
          },
          "is_verified": {
            "type": "boolean",
            "title": "Is Verified",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "document_name",
          "chunk_text",
          "relevance_score"
        ],
        "title": "Citation"
      },
      "ExecutionTrace": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "intent": {
            "type": "string",
            "title": "Intent"
          },
          "planning_steps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Planning Steps"
          },
          "tool_executions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Tool Executions"
          },
          "retrieved_citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Retrieved Citations"
          },
          "rag_pipeline": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RAGPipelineTrace"
              },
              {
                "type": "null"
              }
            ]
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          }
        },
        "type": "object",
        "required": [
          "id",
          "conversation_id",
          "agent_id",
          "intent",
          "planning_steps",
          "tool_executions",
          "retrieved_citations",
          "latency_ms"
        ],
        "title": "ExecutionTrace"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KnowledgeIngestRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Document title"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Raw text content to chunk and index"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "title": "KnowledgeIngestRequest"
      },
      "RAGPipelineTrace": {
        "properties": {
          "query_understanding": {
            "additionalProperties": true,
            "type": "object",
            "title": "Query Understanding"
          },
          "query_rewrite": {
            "additionalProperties": true,
            "type": "object",
            "title": "Query Rewrite"
          },
          "hybrid_retrieval": {
            "additionalProperties": true,
            "type": "object",
            "title": "Hybrid Retrieval"
          },
          "rrf_fusion": {
            "additionalProperties": true,
            "type": "object",
            "title": "Rrf Fusion"
          },
          "reranking": {
            "additionalProperties": true,
            "type": "object",
            "title": "Reranking"
          },
          "context_assembly": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context Assembly"
          },
          "grounding_verification": {
            "additionalProperties": true,
            "type": "object",
            "title": "Grounding Verification"
          }
        },
        "type": "object",
        "required": [
          "query_understanding",
          "query_rewrite",
          "hybrid_retrieval",
          "rrf_fusion",
          "reranking",
          "context_assembly",
          "grounding_verification"
        ],
        "title": "RAGPipelineTrace"
      },
      "RAGQueryRequest": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question",
            "description": "Query for knowledge retrieval"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          },
          "top_k": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K",
            "default": 3
          },
          "min_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Score",
            "default": 0.2
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "RAGQueryRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__models__ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "User question or prompt"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          },
          "customer_identifier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Identifier",
            "default": "guest_user"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel",
            "default": "PLAYGROUND"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest"
      },
      "app__models__ChatResponse": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "response": {
            "type": "string",
            "title": "Response"
          },
          "interactive_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interactive Payload"
          },
          "trace": {
            "$ref": "#/components/schemas/ExecutionTrace"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "message_id",
          "response",
          "trace"
        ],
        "title": "ChatResponse"
      }
    }
  }
}